Refactor PV scheduling library into separate package
