+++ pkg/controller/volume/persistentvolume/index.go
@@ -21,13 +21,9 @@ import (
 	"sort"
 
 	"k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/resource"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/labels"
-	utilfeature "k8s.io/apiserver/pkg/util/feature"
 	"k8s.io/client-go/tools/cache"
 	v1helper "k8s.io/kubernetes/pkg/apis/core/v1/helper"
-	"k8s.io/kubernetes/pkg/features"
+	pvutil "k8s.io/kubernetes/pkg/controller/volume/persistentvolume/util"
 	volumeutil "k8s.io/kubernetes/pkg/volume/util"
 )
 
@@ -96,7 +92,7 @@ func (pvIndex *persistentVolumeOrderedIndex) findByClaim(claim *v1.PersistentVol
 			return nil, err
 		}
 
-		bestVol, err := findMatchingVolume(claim, volumes, nil /* node for topology binding*/, nil /* exclusion map */, delayBinding)
+		bestVol, err := pvutil.FindMatchingVolume(claim, volumes, nil /* node for topology binding*/, nil /* exclusion map */, delayBinding)
 		if err != nil {
 			return nil, err
 		}
@@ -106,176 +102,6 @@ func (pvIndex *persistentVolumeOrderedIndex) findByClaim(claim *v1.PersistentVol
 		}
 	}
 	return nil, nil
-}
-
-// findMatchingVolume goes through the list of volumes to find the best matching volume
-// for the claim.
-//
-// This function is used by both the PV controller and scheduler.
-//
-// delayBinding is true only in the PV controller path.  When set, prebound PVs are still returned
-// as a match for the claim, but unbound PVs are skipped.
-//
-// node is set only in the scheduler path. When set, the PV node affinity is checked against
-// the node's labels.
-//
-// excludedVolumes is only used in the scheduler path, and is needed for evaluating multiple
-// unbound PVCs for a single Pod at one time.  As each PVC finds a matching PV, the chosen
-// PV needs to be excluded from future matching.
-func findMatchingVolume(
-	claim *v1.PersistentVolumeClaim,
-	volumes []*v1.PersistentVolume,
-	node *v1.Node,
-	excludedVolumes map[string]*v1.PersistentVolume,
-	delayBinding bool) (*v1.PersistentVolume, error) {
-
-	var smallestVolume *v1.PersistentVolume
-	var smallestVolumeQty resource.Quantity
-	requestedQty := claim.Spec.Resources.Requests[v1.ResourceName(v1.ResourceStorage)]
-	requestedClass := v1helper.GetPersistentVolumeClaimClass(claim)
-
-	var selector labels.Selector
-	if claim.Spec.Selector != nil {
-		internalSelector, err := metav1.LabelSelectorAsSelector(claim.Spec.Selector)
-		if err != nil {
-			// should be unreachable code due to validation
-			return nil, fmt.Errorf("error creating internal label selector for claim: %v: %v", claimToClaimKey(claim), err)
-		}
-		selector = internalSelector
-	}
-
-	// Go through all available volumes with two goals:
-	// - find a volume that is either pre-bound by user or dynamically
-	//   provisioned for this claim. Because of this we need to loop through
-	//   all volumes.
-	// - find the smallest matching one if there is no volume pre-bound to
-	//   the claim.
-	for _, volume := range volumes {
-		if _, ok := excludedVolumes[volume.Name]; ok {
-			// Skip volumes in the excluded list
-			continue
-		}
-
-		volumeQty := volume.Spec.Capacity[v1.ResourceStorage]
-
-		// check if volumeModes do not match (feature gate protected)
-		isMismatch, err := checkVolumeModeMismatches(&claim.Spec, &volume.Spec)
-		if err != nil {
-			return nil, fmt.Errorf("error checking if volumeMode was a mismatch: %v", err)
-		}
-		// filter out mismatching volumeModes
-		if isMismatch {
-			continue
-		}
-
-		// check if PV's DeletionTimeStamp is set, if so, skip this volume.
-		if utilfeature.DefaultFeatureGate.Enabled(features.StorageObjectInUseProtection) {
-			if volume.ObjectMeta.DeletionTimestamp != nil {
-				continue
-			}
-		}
-
-		nodeAffinityValid := true
-		if node != nil {
-			// Scheduler path, check that the PV NodeAffinity
-			// is satisfied by the node
-			err := volumeutil.CheckNodeAffinity(volume, node.Labels)
-			if err != nil {
-				nodeAffinityValid = false
-			}
-		}
-
-		if IsVolumeBoundToClaim(volume, claim) {
-			// this claim and volume are pre-bound; return
-			// the volume if the size request is satisfied,
-			// otherwise continue searching for a match
-			if volumeQty.Cmp(requestedQty) < 0 {
-				continue
-			}
-
-			// If PV node affinity is invalid, return no match.
-			// This means the prebound PV (and therefore PVC)
-			// is not suitable for this node.
-			if !nodeAffinityValid {
-				return nil, nil
-			}
-
-			return volume, nil
-		}
-
-		if node == nil && delayBinding {
-			// PV controller does not bind this claim.
-			// Scheduler will handle binding unbound volumes
-			// Scheduler path will have node != nil
-			continue
-		}
-
-		// filter out:
-		// - volumes in non-available phase
-		// - volumes bound to another claim
-		// - volumes whose labels don't match the claim's selector, if specified
-		// - volumes in Class that is not requested
-		// - volumes whose NodeAffinity does not match the node
-		if volume.Status.Phase != v1.VolumeAvailable {
-			// We ignore volumes in non-available phase, because volumes that
-			// satisfies matching criteria will be updated to available, binding
-			// them now has high chance of encountering unnecessary failures
-			// due to API conflicts.
-			continue
-		} else if volume.Spec.ClaimRef != nil {
-			continue
-		} else if selector != nil && !selector.Matches(labels.Set(volume.Labels)) {
-			continue
-		}
-		if v1helper.GetPersistentVolumeClass(volume) != requestedClass {
-			continue
-		}
-		if !nodeAffinityValid {
-			continue
-		}
-
-		if node != nil {
-			// Scheduler path
-			// Check that the access modes match
-			if !checkAccessModes(claim, volume) {
-				continue
-			}
-		}
-
-		if volumeQty.Cmp(requestedQty) >= 0 {
-			if smallestVolume == nil || smallestVolumeQty.Cmp(volumeQty) > 0 {
-				smallestVolume = volume
-				smallestVolumeQty = volumeQty
-			}
-		}
-	}
-
-	if smallestVolume != nil {
-		// Found a matching volume
-		return smallestVolume, nil
-	}
-
-	return nil, nil
-}
-
-// checkVolumeModeMismatches is a convenience method that checks volumeMode for PersistentVolume
-// and PersistentVolumeClaims
-func checkVolumeModeMismatches(pvcSpec *v1.PersistentVolumeClaimSpec, pvSpec *v1.PersistentVolumeSpec) (bool, error) {
-	if !utilfeature.DefaultFeatureGate.Enabled(features.BlockVolume) {
-		return false, nil
-	}
-
-	// In HA upgrades, we cannot guarantee that the apiserver is on a version >= controller-manager.
-	// So we default a nil volumeMode to filesystem
-	requestedVolumeMode := v1.PersistentVolumeFilesystem
-	if pvcSpec.VolumeMode != nil {
-		requestedVolumeMode = *pvcSpec.VolumeMode
-	}
-	pvVolumeMode := v1.PersistentVolumeFilesystem
-	if pvSpec.VolumeMode != nil {
-		pvVolumeMode = *pvSpec.VolumeMode
-	}
-	return requestedVolumeMode != pvVolumeMode, nil
 }
 
 // findBestMatchForClaim is a convenience method that finds a volume by the claim's AccessModes and requests for Storage
@@ -361,20 +187,4 @@ func claimToClaimKey(claim *v1.PersistentVolumeClaim) string {
 
 func claimrefToClaimKey(claimref *v1.ObjectReference) string {
 	return fmt.Sprintf("%s/%s", claimref.Namespace, claimref.Name)
-}
-
-// Returns true if PV satisfies all the PVC's requested AccessModes
-func checkAccessModes(claim *v1.PersistentVolumeClaim, volume *v1.PersistentVolume) bool {
-	pvModesMap := map[v1.PersistentVolumeAccessMode]bool{}
-	for _, mode := range volume.Spec.AccessModes {
-		pvModesMap[mode] = true
-	}
-
-	for _, mode := range claim.Spec.AccessModes {
-		_, ok := pvModesMap[mode]
-		if !ok {
-			return false
-		}
-	}
-	return true
 }
