+++ pkg/controller/volume/scheduling/scheduler_bind_cache_metrics.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package persistentvolume
+package scheduling
 
 import (
 	"github.com/prometheus/client_golang/prometheus"
