+++ pkg/controller/volume/scheduling/scheduler_binder_fake.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package persistentvolume
+package scheduling
 
 import "k8s.io/api/core/v1"
